+++ src/main/kotlin/com/coder/toolbox/sdk/ex/APIResponseException.kt
@@ -10,6 +10,7 @@ class APIResponseException(action: String, url: URL, code: Int, errorResponse: A
 
     val reason = errorResponse?.detail
     val isUnauthorized = HttpURLConnection.HTTP_UNAUTHORIZED == code
+    val isTokenExpired = isUnauthorized && reason?.contains("API key expired") == true
 
     companion object {
         private fun formatToPretty(